requirements.txt
Package             Version
------------------- ---------
bleach              6.0.0
bokeh               3.1.1
certifi             2023.7.22
charset-normalizer  3.2.0
colorama            0.4.6
contourpy           1.1.0
cycler              0.11.0
fonttools           4.42.1
idna                3.4
importlib-metadata  6.8.0
importlib-resources 6.0.1
Jinja2              3.1.2
kafka-python        2.0.2
kiwisolver          1.4.5
linkify-it-py       2.0.2
Markdown            3.4.4
markdown-it-py      3.0.0
MarkupSafe          2.1.3
matplotlib          3.7.2
mdit-py-plugins     0.4.0
mdurl               0.1.2
numpy               1.24.4
packaging           23.1
pandas              2.0.3
panel               1.2.2
param               1.13.0
Pillow              10.0.0
pip                 23.2.1
pyparsing           3.0.9
python-dateutil     2.8.2
pytz                2023.3
pyviz_comms         3.0.0
PyYAML              6.0.1
requests            2.31.0
setuptools          68.1.2
six                 1.16.0
tornado             6.3.3
tqdm                4.66.1
typing_extensions   4.7.1
tzdata              2023.3
uc-micro-py         1.0.2
urllib3             2.0.4
webencodings        0.5.1
wheel               0.41.2
xyzservices         2023.7.0
zipp                3.16.2
